requirement.txt
flask

flask-sqlalchemy

gunicorn

pytest
pytest-flask
pytest-cov
